desktop_agent/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
pywin32>=306
mss>=9.0.0
psutil>=5.9.0
GPUtil>=1.4.0
pycaw>=20230407
comtypes>=1.2.0
pyperclip>=1.8.0
plyer>=2.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
structlog>=23.2.0
Pillow>=10.0.0
